+++ xlfg/__init__.py
@@ -8,4 +8,4 @@ from __future__ import annotations
 
 __all__ = ["__version__"]
 
-__version__ = "2.0.5"
+__version__ = "2.0.6"